#9751: Restructure ttnn transformers to new folder structure
